candidate_service/requirements.txt
mcp-server>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
arcadepy==1.3.0
openinference-instrumentation-litellm==0.1.19
arize==7.41.1
litellm==1.70.0 
arize-otel==0.8.1 
python-dotenv==1.1.0
httpx==0.27.0 
groq==0.25.0 